// featurehub-javascript-client-sdk/app/models/models/feature-state.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.8
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FeatureRolloutStrategy } from "./feature-rollout-strategy";
import { FeatureValueType } from "./feature-value-type";

/**
 *
 * @export
 * @interface FeatureState
 */
export interface FeatureState {
  /**
   *
   * @type {string}
   * @memberof FeatureState
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof FeatureState
   */
  key: string;
  /**
   * Is this feature locked. Usually this doesn\'t matter because the value is the value, but for FeatureInterceptors it can matter.
   * @type {boolean}
   * @memberof FeatureState
   */
  l?: boolean;
  /**
   * The version of the feature, this allows features to change values and it means we don\'t trigger events
   * @type {number}
   * @memberof FeatureState
   */
  version?: number;
  /**
   *
   * @type {FeatureValueType}
   * @memberof FeatureState
   */
  type?: FeatureValueType;
  /**
   * the current value
   * @type {any}
   * @memberof FeatureState
   */
  value?: any | null;
  /**
   * strategy id that was applied if this was a server evaluated feature
   * @type {string}
   * @memberof FeatureState
   */
  v?: string | null;
  /**
   * Any additional properties configured by System Operator
   * @type {Record<string, string>}
   * @memberof FeatureState
   */
  fp?: Record<string, string> | null;
  /**
   * This field is filled in from the client side in the GET api as the GET api is able to request multiple environments. It is never passed from the server, as an array of feature states is wrapped in an environment.
   * @type {string}
   * @memberof FeatureState
   */
  environmentId?: string | null;
  /**
   *
   * @type {Array<FeatureRolloutStrategy>}
   * @memberof FeatureState
   */
  strategies?: Array<FeatureRolloutStrategy> | null;
}